optimization v2

+++ moves.cpp
@@ -1,6 +1,7 @@
 #include "moves.hpp"
+#include <cstdint>
 
-Move::Move(int from,int to,int flags){
+Move::Move(uint8_t from, uint8_t to, uint8_t flags){
     this->from = from;
     this->to = to;
     this->flags = flags;
