+++ moves.hpp
@@ -1,11 +1,12 @@
 #ifndef MOVES_HPP
 #define MOVES_HPP
 
+#include <cstdint>
 class Move{
     public:
-        int from;
-        int to;
-        int flags;
-        Move(int from,int to,int flags);
+        uint8_t from;
+        uint8_t to;
+        uint8_t flags;
+        Move(uint8_t from, uint8_t to, uint8_t flags);
 };
 #endif
